Tutorial - create an Azure Function to track the changes in CosmosDB
Change feed in Azure Cosmos DB is a persistent record of changes to a container in the order they occur. Change feed support in Azure Cosmos DB works by listening to an Azure Cosmos container for any changes. It then outputs the sorted list of documents that were changed in the order in which they were modified.
In this tutorial, I will show how to set up the Azure Function that will track the item changes to the CosmosDB container.